+++ modules/arr-common/mkArrServiceModule.nix
@@ -51,6 +51,7 @@ in
     (import ./mediaManagement.nix { inherit serviceName; })
     (import ./mediaDirs.nix { inherit serviceName; })
     (import ./postgres.nix { inherit serviceName; })
+    (import ./releaseProfiles.nix { inherit serviceName; })
     (import ./rootFolders.nix { inherit serviceName; })
     (import ./remotePathMappings.nix { inherit serviceName; })
   ];